frontend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
uvicorn[standard]==0.23.2
channels==4.0.0
channels-redis==4.1.0
whitenoise==6.6.0
paho-mqtt==1.6.1
